refactor(types): extract shared types to common module and remove duplicates
- Create types/common.ts with shared primitive types (Timestamp, Metadata, etc.)
- Remove duplicated type definitions from timeSeries/types.ts
- Import shared types from common module instead of duplicating
- Replace inline import() syntax with top-level import for TimeseriesFilter
- Update syntheticTimeSeriesApi.ts to import DatapointInfo from common module
