+++ packages/stable/src/api/timeSeries/syntheticTimeSeriesApi.ts
@@ -1,7 +1,7 @@
 // Copyright 2020 Cognite AS
 
 import { BaseResourceAPI } from '@cognite/sdk-core';
-import type { DatapointInfo } from '../../types';
+import type { DatapointInfo } from '../../types/common';
 import type { SyntheticQuery, SyntheticQueryResponse } from './types';
 
 export class SyntheticTimeSeriesAPI extends BaseResourceAPI<SyntheticQueryResponse> {
